src/requirements.txt
pynacl
requests
boto3